+++ src/__core/neeko.ts
@@ -8,27 +8,45 @@ import {
 } from 'mobx'
 
 import { getInstance } from './di'
-import { isGenerator, assertKeyValid } from '../__internal'
+import { isGenerator, assertKeyValid, throwError } from '../__internal'
 import { plugins } from './plugin'
 import { model as modelFn } from '../types'
 
 // okeen generate instance
-function getInstanceByOkeen(fn: <S>(ins: S) => S) {
+function getInstanceByOkeen(
+  // add key in instance of Neeko
+  transferIns: <S>(ins: S) => S,
+  // add method in Neeko
+  transferProto: <P>(proto: P) => P,
+) {
   function getInsByModel() {
-    let ins = {}
-    /* istanbul ignore else */
-    if (typeof fn === 'function') {
-      ins = fn(ins)
-    }
+    const Neeko = function () {}
+
+    transferProto?.(Neeko.prototype)
+    Neeko.prototype.$new = () => getInstanceByOkeen(transferIns, transferProto)
+
+    // @ts-ignore
+    let ins = new Neeko()
+
+    // bind ins on each Neeko.prototype
+    Object.keys(Neeko.prototype).forEach((method) => {
+      const boundedMethod = Neeko.prototype[method].bind(ins)
+
+      // copy static on bounded method, eg. __isEffect
+      Object.keys(Neeko.prototype[method]).forEach((key) => {
+        boundedMethod[key] = Neeko.prototype[method][key]
+      })
+
+      Neeko.prototype[method] = boundedMethod
+    })
+
+    ins = transferIns?.(ins)
 
     return ins
   }
 
   // use {} for uid
   const ins = getInstance<any>({}, getInsByModel)
-
-  ins.$new = () => getInstanceByOkeen(fn)
-
   return ins
 }
 
@@ -69,12 +87,14 @@ function transferComputed(ins: any, _computed: any = {}) {
   })
 }
 
-let updateTask: any[] = []
-
 // the only method(like reducer) to change state is $update
-function transferReducers(ins: any) {
+function transferReducers(proto: any) {
+  let updateTask: any[] = []
   // add default $update method
-  const update = async (stateOrUpdater: any, nextTick: boolean = false) => {
+  const update = function (stateOrUpdater: any, nextTick: boolean = false) {
+    // @ts-ignore
+    const ins = this
+
     // 1. prepare new state
     let state: any = {}
     if (typeof stateOrUpdater === 'function') {
@@ -102,24 +122,22 @@ function transferReducers(ins: any) {
       state = stateOrUpdater
     }
 
+    // 1.5 precheck error
+    for (const key in state) {
+      if (!isObservableProp(ins, key)) {
+        throwError(`cannot update key (${key}) not in state`)
+      }
+    }
+
     // 2. update state
     const task = () => {
       for (const key in state) {
-        if (isObservableProp(ins, key)) {
-          ins[key] = state[key]
-        } else {
-          console.error(`[okeen]: cannot update key (${key}) not in state`)
-          // extendObservable(ins, { [key]: state[key] });
-        }
+        // no need check isObservableProp because 1.5 has checked
+        ins[key] = state[key]
       }
     }
 
     updateTask.push(task)
-
-    // nextTick
-    if (nextTick) {
-      await Promise.resolve()
-    }
 
     // sync $update will complete all async tasks
     const runAllTask = () => {
@@ -130,11 +148,15 @@ function transferReducers(ins: any) {
     }
 
     const taskWithAction = action(runAllTask.bind(ins))
-    taskWithAction()
-    return
-  }
 
-  ins.$update = update
+    if (nextTick) {
+      // nextTick
+      Promise.resolve().then(() => taskWithAction())
+    } else {
+      taskWithAction()
+    }
+  }
+  proto.$update = update
 }
 
 export function isEffect(fn: (args: any[]) => any) {
@@ -143,42 +165,57 @@ export function isEffect(fn: (args: any[]) => any) {
 }
 
 export function setEffect(ins: any, key: string, fn: (args: any[]) => any) {
-  fn = fn.bind(ins)
-  if (isGenerator(fn)) {
-    // TODO: remove generator flow
-    ins[key] = flow(fn)
-  } else {
-    ins[key] = fn
-  }
-
-  ins[key].__isEffect = true
+  const fixedFn = fn
+  // @ts-ignore
+  fixedFn.__isEffect = true
+  Object.setPrototypeOf(ins, {
+    [key]: fn,
+  })
 }
 
-function transferEffects(ins: any, _effects: any = {}) {
-  bindOptions(ins, _effects, (key) => {
-    setEffect(ins, key, _effects[key])
+export function setEffectProto(
+  proto: any,
+  key: string,
+  fn: (args: any[]) => any,
+) {
+  const fixedFn = isGenerator(fn) ? flow(fn) : fn
+  // @ts-ignore
+  fixedFn.__isEffect = true
+
+  proto[key] = fixedFn
+}
+
+function transferEffects(proto: any, _effects: any = {}) {
+  bindOptions(proto, _effects, (key) => {
+    setEffectProto(proto, key, _effects[key])
   })
 }
 
 const model: typeof modelFn = (options: any) => {
-  const ins = getInstanceByOkeen((ins) => {
-    transferReducers(ins)
-    transferState(ins, options.state)
-    transferComputed(ins, options.computed)
-    transferEffects(ins, options.effects)
-    transferRef(ins, options.ref)
+  const ins = getInstanceByOkeen(
+    (ins) => {
+      transferState(ins, options.state)
+      transferComputed(ins, options.computed)
+      transferRef(ins, options.ref)
 
-    // register plugins after transfer
-    plugins.forEach(({ key, type, fn }) => {
-      if (type === 'internal') {
-        fn(ins, options[key])
-      } else {
-        fn(ins, options.plugins && options.plugins[key])
-      }
-    })
+      // register plugins after transfer
+      plugins.forEach(({ key, type, fn }) => {
+        if (type === 'internal') {
+          fn(ins, options[key])
+        } else {
+          fn(ins, options.plugins && options.plugins[key])
+        }
+      })
 
-    return ins
-  })
+      return ins
+    },
+    (proto) => {
+      transferReducers(proto)
+      transferEffects(proto, options.effects)
+
+      return proto
+    },
+  )
 
   return ins as any
 }